backend/requirements.txt
# Course Copilot Backend

fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-generativeai==0.3.2
pypdf2==3.0.1
pandas==2.2.0
python-dotenv==1.0.0
pydantic[email]==2.6.0
alembic==1.13.1
